Upload Profile Picutre For Customer Added

+++ src/Zolupos.Server/Zolupos.Application/Common/DTO/PaymentDTO.cs
@@ -5,8 +5,8 @@ namespace Zolupos.Application.Common.DTO
     public class PaymentDTO
     {
 		public int PaymentType { get; set; }
-        public int Tendered { get; set; }
-        public int Change { get; set; }
-        public int Amount { get; set; }
+        public float Tendered { get; set; }
+        public float Change { get; set; }
+        public float Amount { get; set; }
     }
 }
